% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oviz-package.R
\docType{data}
\name{ionwaters}
\alias{ionwaters}
\title{Ion concentrations in different water samples}
\format{a \code{data.frame}. The concentrations in mgL^-1 are given for eight ions: K, Na, Ca, Mg, Cl, SO4, HCO3, CO3
sampled in six places: Adriatic Sea, lake Balaton, lake Davos, Lower Lunz lake, Aggtelek underground water and Szomod spring.}
\source{
Maucha, R. Hydrochemische Methoden in der Limnologie: mit besonderer Ber\"ucksichtigung
der Verfahren von LW Winkler. Schweizerbart, 1932. Table 14 pp. 86.
}
\usage{
data(ionwaters)
}
\description{
This dataset gives the concentration of the eight main ions in six type of natural waters.
}
\keyword{datasets}

