% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maucha.R
\name{maucha}
\alias{maucha}
\title{Maucha diagram}
\usage{
maucha(x, col = NULL, labels = TRUE, labels.col = 1, polygon = FALSE,
  polygon.col = 1, main = NULL, ...)
}
\arguments{
\item{x}{a vector or a matrix giving the concentrations of the ions in mgL^-1.
The values must be ordered as K, Na, Ca, Mg, SO4, Cl, HCO3, CO3.}

\item{col}{a vector of R colors for each ion. Recycled if necessary.}

\item{labels}{a logical stating whether labels are displayed.}

\item{labels.col}{a vector of R colors for each label. Recycled if necessary.}

\item{polygon}{a logical stating whether the hexadecagon is drawn.}

\item{polygon.col}{a vector of R colors for the lines of each of the eight sector of the hexadecagon.}

\item{main}{a character string giving a main title for the diagrams.}

\item{...}{other arguments to be passed to the \code{\link[graphics]{plot.default}} function.}
}
\description{
This function draws Maucha diagram to represent the main ion composition of a water sample.
}
\references{
Maucha, R. Hydrochemische Methoden in der Limnologie: mit besonderer Berücksichtigung
der Verfahren von LW Winkler. Schweizerbart, 1932.
}

